Remove the useless StreamIdByte property and simplify logging logic
This also adds a few checks and pieces of info in headers to `/logs/getStream`

+++ Sofa_API/Program.cs
@@ -77,7 +77,7 @@ if (DockerLocal.IsDockerComposeAvailable)
 
 builder.Logging.ClearProviders();
 var app = builder.Build();
-if (ApiConfig.TerminalVerbosity > (byte) StreamId.Request || ApiConfig.ApiConfiguration.LogVerbosity > (byte) StreamId.Request)
+if (ApiConfig.TerminalVerbosity > StreamId.Request || ApiConfig.ApiConfiguration.LogVerbosity > StreamId.Request)
 	app.UseMiddleware<RequestLoggingMiddleware>();
 
 if (Environment.OSVersion.Platform != PlatformID.Unix)
